modify

+++ lightning_attn_cute/test.py
@@ -81,22 +81,22 @@ def lightning_attn_no_decay(
     for i in range(NUM_BLOCK):
         si = i * BLOCK
         ei = min(si + BLOCK, N)
-        qi = q[:, :, si:ei, :].contiguous().to(torch.float32)
-        ki = k[:, :, si:ei, :].contiguous().to(torch.float32)
-        vi = v[:, :, si:ei, :].contiguous().to(torch.float32)
+        qi = q[:, :, si:ei, :].contiguous().to(torch.float16)
+        ki = k[:, :, si:ei, :].contiguous().to(torch.float16)
+        vi = v[:, :, si:ei, :].contiguous().to(torch.float16)
 
-        qkv_none_diag = torch.matmul(qi, kv.to(qi.dtype))
+        qkv_none_diag = torch.matmul(qi, kv.to(qi.dtype)).to(torch.float32)
         o_inter_output[i] = qkv_none_diag.detach().clone()
 
         # diag
         qk = torch.matmul(qi, ki.transpose(-1, -2))
 
-        qkv_diag = torch.matmul(qk, vi)
+        qkv_diag = torch.matmul(qk, vi).to(torch.float32)
         o_intra_output[i] = qkv_diag.detach().clone()
 
         output[:, :, si:ei] = (qkv_none_diag + qkv_diag).to(torch.float16)
         # new_kv = torch.matmul(ki.transpose(-1, -2).to(vi.dtype), vi).to(torch.float32)
-        new_kv = torch.matmul(ki.transpose(-1, -2).to(vi.dtype), vi)
+        new_kv = torch.matmul(ki.transpose(-1, -2), vi).to(torch.float32)
         kv = kv + new_kv
         kv_output[i] = kv.detach().clone()
